Land Claim Kit
I increased the distance of the land claim kit. I noticed, and it was brought up by the guys at the other base, they wanted to cover a larger area to account for their walls, but you can't put multiple land claim kits down if they overlap each other. So I though to rectify this we could increase the radius. It was set to 80 prior to moving to this map but I didn't want to make it too big.