Update TraderPlusPriceConfig.json
I added the items to the stash sections, but didn't add them to the trader. There is something wrong. Prior to doing this I noticed items that were in the stash but couldn't be sold, they say -1 when you go to sell. I though it was a code issue but all the numbers are identical. So the Items are added but some still can't be sold to the stash....see any of the JMC items